Migrate the Meter Structs to dc_bundle
Fixes #1452
